Virtually finished code.
The only problem is that it takes several attempts to push the button for it to recognize and either pause the timer, or stop the buzzer. This is because we use the delay() function, during which no input can be processed, so you have to happen to push the button in between one of these states. If we can find a way for it to delay and still be open to input, that would be better.